# Conflicts:
#	hailo_apps/config/config.yaml
#	hailo_apps/postprocess/cpp/meson.build

+++ hailo_apps/cpp/instance_segmentation/build.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
+set -e
 
-cmake -H. -Bbuild -DCMAKE_RUNTIME_OUTPUT_DIRECTORY=$PWD/build/x86_64
-cmake --build build
+mkdir -p build
+cmake -H. -Bbuild -DCMAKE_BUILD_TYPE=Release
+cmake --build build -- -j"$(nproc)"
 
-if [[ -f "hailort.log" ]]; then
-    rm hailort.log
-fi
+[[ -f hailort.log ]] && rm hailort.log
